+++ src/main/java/coupon/exception/CouponException.java
@@ -1,4 +1,4 @@
-package coupon.domain.exception;
+package coupon.exception;
 
 public class CouponException extends RuntimeException {
     public CouponException() {
