+++ src/main/java/coupon/exception/CouponNameValidationException.java
@@ -1,4 +1,4 @@
-package coupon.domain.exception;
+package coupon.exception;
 
 public class CouponNameValidationException extends CouponException {
     public CouponNameValidationException() {
